+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 [Downloadable releases](https://github.com/Keplerjs/Kepler/releases)
 
 ## Next version
+- new user field loginIp
 - plugin admin new sessions section in user admin panel
 - custom meta and link tags by settings by K.settings.public.router.meta
 - plugin admin create route for each loaded places and users, fix categories tagsinput bug